Production/Back_End/requirements.txt
click==7.1.2
DateTime==4.3
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
zope.interface==5.1.0
Flask-Cors==1.10.3
Flask-Caching==1.9.0
redis==3.5.3
